organise methods in genric instrument class, keithley class and k2400 class
